backend/requirements.txt
Django>=4.2,<5.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
google-generativeai>=0.8
gTTS>=2.5
channels>=4.1
